run serial jobs in parallel TACC launcher

+++ inun_launcher_prep.sh
@@ -3,10 +3,13 @@
 # make subdirectories for inudnation .tif files
 # create file of launcher commands
 
-rm inun_launcher.sh
+if [ -f inun_launcher.sh ]; then
+    rm inun_launcher.sh
+fi
+
 for D in 12*; do
     if [ -d "${D}" ]; then
-        mkdir -p ${D}/inundation/mod{1,2a,2b,3a,3b}
+        mkdir -p ${D}/inundation_rev/mod{1,2a,2b,3a,3b}
         echo "python calc_inun.py \"${D}\"" >> inun_launcher.sh
     fi
 done
